// client/src/generated/GithubPrSearchItem.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { GithubUserRef } from "./GithubUserRef";

/**
 * A single PR item returned by `GET /api/v1/github/prs/`.
 */
export type GithubPrSearchItem = { number: number, title: string, html_url: string, repository_url: string, updated_at: string, user: GithubUserRef, };